+++ src/dune_file.ml
@@ -333,9 +333,9 @@ module Blang = struct
     decode
 end
 
-let enabled_if =
+let enabled_if ~since =
   field "enabled_if" ~default:Blang.true_
-    (Syntax.since Stanza.syntax (1, 4) >>> Blang.decode)
+    (Syntax.since Stanza.syntax since >>> Blang.decode)
 
 module Per_module = struct
   include Per_item.Make(Module.Name)
@@ -823,6 +823,7 @@ module Library = struct
     ; implements               : (Loc.t * Lib_name.t) option
     ; private_modules          : Ordered_set_lang.t option
     ; stdlib                   : Stdlib.t option
+    ; enabled_if               : Blang.t
     }
 
   let decode =
@@ -871,6 +872,8 @@ module Library = struct
            >>= fun () -> Ordered_set_lang.decode)
        and stdlib =
          field_o "stdlib" (Syntax.since Stdlib.syntax (0, 1) >>> Stdlib.decode)
+       and enabled_if =
+         enabled_if ~since:(1, 7)
        in
        let wrapped = Wrapped.make ~wrapped ~implements in
        let name =
@@ -925,6 +928,14 @@ module Library = struct
              "A library cannot use (self_build_stubs_archive ...) \
               and (%s ...) simultaneously." name
        in
+       Blang.fold_vars enabled_if ~init:() ~f:(fun var () ->
+         match String_with_vars.Var.name var,
+               String_with_vars.Var.payload var with
+         | "os_type", None -> ()
+         | _ ->
+           Errors.fail (String_with_vars.Var.loc var)
+             "Only the 'os_type' variable is allowed in the 'enabled_if' \
+              field of libraries.");
        { name
        ; public
        ; synopsis
@@ -952,6 +963,7 @@ module Library = struct
        ; implements
        ; private_modules
        ; stdlib
+       ; enabled_if
        })
 
   let has_stubs t =
@@ -1444,7 +1456,7 @@ module Rule = struct
           | false, Some mode -> Ok mode
           | true, None -> Ok Fallback
           | false, None -> Ok Standard)
-    and enabled_if = enabled_if
+    and enabled_if = enabled_if ~since:(1, 4)
     in
     { targets
     ; deps
@@ -1517,7 +1529,7 @@ module Rule = struct
             record
               (let%map modules = field "modules" (list string)
                and mode = Mode.field
-               and enabled_if = enabled_if
+               and enabled_if = enabled_if ~since:(1, 4)
                in
                { modules; mode; enabled_if }))
           ~else_:(
@@ -1603,7 +1615,7 @@ module Menhir = struct
        and mode = Rule.Mode.field
        and infer = field_o_b "infer" ~check:(Syntax.since syntax (2, 0))
        and menhir_syntax = Syntax.get_exn syntax
-       and enabled_if = enabled_if
+       and enabled_if = enabled_if ~since:(1, 4)
        in
        let infer =
          match infer with
@@ -1702,7 +1714,7 @@ module Tests = struct
                      ~default:Executables.Link_mode.Set.default
        and deps =
          field "deps" (Bindings.decode Dep_conf.decode) ~default:Bindings.empty
-       and enabled_if = enabled_if
+       and enabled_if = enabled_if ~since:(1, 4)
        and action =
          field_o
            "action"