+++ src/lib.mli
@@ -236,6 +236,7 @@ module DB : sig
     -> has_native:bool
     -> ext_lib:string
     -> ext_obj:string
+    -> os_type:string
     -> (Path.t * Dune_file.Library.t) list
     -> t